+++ minimalkv/net/botostore.py
@@ -1,8 +1,7 @@
 from contextlib import contextmanager
-from typing import Dict, Iterator
+from typing import IO, Dict, Iterator, cast
 
 from minimalkv import CopyMixin, KeyValueStore, UrlMixin
-from minimalkv._typing import File
 
 
 @contextmanager
@@ -99,7 +98,7 @@ class BotoStore(KeyValueStore, UrlMixin, CopyMixin):  # noqa D
         with map_boto_exceptions(key=key):
             return k.get_contents_as_string()
 
-    def _get_file(self, key: str, file: File) -> str:
+    def _get_file(self, key: str, file: IO) -> str:
         k = self.__new_key(key)
         with map_boto_exceptions(key=key):
             return k.get_contents_to_file(file)
@@ -109,7 +108,7 @@ class BotoStore(KeyValueStore, UrlMixin, CopyMixin):  # noqa D
         with map_boto_exceptions(key=key):
             return k.get_contents_to_filename(filename)
 
-    def _open(self, key: str) -> File:
+    def _open(self, key: str) -> IO:
         from boto.s3.keyfile import KeyFile
 
         class SimpleKeyFile(KeyFile):  # noqa D
@@ -128,7 +127,7 @@ class BotoStore(KeyValueStore, UrlMixin, CopyMixin):  # noqa D
 
         k = self.__new_key(key)
         with map_boto_exceptions(key=key):
-            return SimpleKeyFile(k)
+            return cast(IO, SimpleKeyFile(k))
 
     def _copy(self, source: str, dest: str) -> None:
         if not self._has_key(source):
@@ -144,7 +143,7 @@ class BotoStore(KeyValueStore, UrlMixin, CopyMixin):  # noqa D
             k.set_contents_from_string(data, **self.__upload_args())
             return key
 
-    def _put_file(self, key: str, file: File) -> str:
+    def _put_file(self, key: str, file: IO) -> str:
         k = self.__new_key(key)
         with map_boto_exceptions(key=key):
             k.set_contents_from_file(file, **self.__upload_args())
