+++ minimalkv/storefact/_boto.py
