+++ minimalkv/storefact/_store_decoration.py
