+++ tests/storefact/test_urls.py
@@ -1,7 +1,8 @@
 import pytest
 
-import minimalkv
-import minimalkv.decorator
+from minimalkv.decorator import ReadOnlyDecorator
+from minimalkv.storefact._urls import url2dict
+from minimalkv.storefact.access import get_store_from_url
 
 good_urls = [
     (
@@ -73,22 +74,22 @@ bad_urls = [
 
 def test_raise_on_invalid_store():
     with pytest.raises(ValueError):
-        minimalkv.url2dict(u"dummy://foo/bar")
+        url2dict(u"dummy://foo/bar")
 
 
 @pytest.mark.parametrize("url, expected", good_urls)
 def test_url2dict(url, expected):
-    assert minimalkv.url2dict(url) == expected
+    assert url2dict(url) == expected
 
 
 @pytest.mark.parametrize("url, raises", bad_urls)
 def test_bad_url2dict(url, raises):
     with pytest.raises(raises):
-        minimalkv.url2dict(url)
+        url2dict(url)
 
 
 def test_roundtrip():
     assert isinstance(
-        minimalkv.get_store_from_url(u"memory://#wrap:readonly"),
-        minimalkv.decorator.ReadOnlyDecorator,
+        get_store_from_url(u"memory://#wrap:readonly"),
+        ReadOnlyDecorator,
     )
